Google Meet updates V1
Added more specifics around google meet v zoom and cleaned this up to be specific to google MEET not hangouts. Didn't change url .md name tho. Removed two screenshots that are obsolete now. I don't actually know if the Available Commands are valid or not still so those might need to be updated or removed.