Update affinity-channels.md
First set of revisions of for our affinity channel policy. I need to do additional research on this. This is a highly complicated issue and touches on state and federal nondiscrimination laws, privacy laws, and labor/union organizing laws. I think we should be very careful about deciding to move forward with officially recognizing these groups. We want to make sure that whatever policy we put in place recognizes and supports all of our employees rights including the rights of the employees in the affinity groups.