Create benefits-and-holidays.md
This is a draft of a Canadian version of our benefits-and-holidays page.  This will set up a Canadian docs area where we can create all necessary documentation that diverges from the US handbook.

Marc - I've left some sections to be filled in once we hear back from PEO Canada, but here are some starting pieces to review.