Equal employment isn't just an obligation
We should state in the handbook that this is something we believe will make us a stronger and more competitive company. The Trinet paragraphs can stay there, but we have to know people (including those we are recruiting) that we aspire to greater things than merely complying with the law.