Update culture.md
Fixing links in 2nd paragraph of the 12th FAQ (2nd from last) that were producing 404 errors. 2nd link is still not resolving to correct section of harassment policy. Need help to fix this! Also changed sentence to start "If you are or know of..." instead of current phrasing "You are or know of...".