Update culture.md (#920)
* Update culture.md

Fixing links in 2nd paragraph of the 12th FAQ (2nd from last) that were producing 404 errors. 2nd link is still not resolving to correct section of harassment policy. Need help to fix this! Also changed sentence to start "If you are or know of..." instead of current phrasing "You are or know of...".

* Update culture.md

Fixed target URLs for 2 harassment policy links.

* Update culture.md

Fixed 2nd link to harassment policy to correct heading.

* Update culture.md

* Update culture.md

Co-authored-by: Daniel Mundra <dmundra@users.noreply.github.com>