Update README.md
Replaced with new content oriented to non-devs making a contribution for the first time, as part of a rewrite of this entire section. Edits were vetted with @dmundra, @grugnog and @kellysmith1008. References to @fa-solid fa-lock will appear as a lock icon when the Font Awesome code is applied to our config file.