Update mkdocs.yml (#1113)
We’ve updated the guidebook top navigation to increase the font color contrast and font size. In doing so the space between the labels is smaller, thus reducing the ability to differentiate between the labels. To correct this, we’re changing the navigation labels using words with equivalent meaning and fewer characters.
"About CivicActions" changes to "About us".
"Common practices and tools" changes to "Our practices and tools".