Update withholds.md
We may want to leave this one out for now -- it has had mixed results in the past, and haven't heard it referenced in several years. I don't know if it feels like a part of the company process or culture at this point but it warrants discussion either way.